api/requirements.txt
fastapi
uvicorn[standard]
pydantic
psycopg2-binary
python-dotenv
osmnx
pandas
joblib
scikit-learn
